+++ Elements/test/VectorTests.cs
@@ -1,4 +1,5 @@
 using Elements.Geometry;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using Xunit;
@@ -568,6 +569,14 @@ namespace Elements.Tests
             Assert.Collection(result,
                 x => x.IsAlmostEqualTo(Vector3.Origin, tolerance),
                 x => x.IsAlmostEqualTo(new Vector3(5, 5), tolerance));
+        }
+
+        [Fact]
+        public void PointsAreCoplanarWithinTolerance()
+        {
+            var ptsSerialized = "[{\"X\":0.0,\"Y\":0.0,\"Z\":0.0},{\"X\":20.0,\"Y\":0.0,\"Z\":-8.43769498715119E-15},{\"X\":19.999999999999996,\"Y\":20.0,\"Z\":-1.021405182655144E-14},{\"X\":9.999999999999995,\"Y\":20.0,\"Z\":-3.096967127191874E-13},{\"X\":10.0,\"Y\":10.0,\"Z\":-4.218847493575595E-15},{\"X\":1.7763568394002505E-15,\"Y\":9.999999999999998,\"Z\":0.0}]";
+            var pts = JsonConvert.DeserializeObject<List<Vector3>>(ptsSerialized);
+            Assert.True(pts.AreCoplanar());
         }
     }
 }