+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Using Multiple `ModelText`s would sometimes result in a corrupted texture atlas, with cutoff text. This is fixed.
 - #865
 - `Network`: intersections are not created for some E-shape cases
+- `Vector3.AreCoplanar` would sometimes return false negatives.
 
 ## 1.6.0